Extract shortcuts for frequently used functionality
